+++ app/Models/User.php
@@ -50,6 +50,7 @@ use Spatie\Permission\Traits\HasRoles;
     'avatar_path',
     'notification_prefs',
     'account_notes',
+    'active_web_session_id',
 ])]
 #[Hidden(['password', 'remember_token'])]
 class User extends Authenticatable implements CanResetPasswordContract